Fix the movement bug
To fix the bug I drastically simplified all math by assuming a fixed
game window size of 1000x1000.
This way, the screen size doesn't change, which was one of the things
complicating the translation math.

From there, the rest of the math also gets drastically simpler as we can just use the
viewport click coordinates instead of translating them.

Afterwards it was just a matter of fixing the simulation code, including
the spawning mechanism.

There's still a React lint going on that I don't quite get, so I'll need
help on that